* Fixed Nginx - Fpm configurations. * Fixed valet share command for port 88 * Replaced mailhog with mailpit
