+++ cli/stubs/init/mailpit
@@ -1,0 +1,11 @@
+[Unit]
+Description=Mailpit
+After=network.target
+
+[Service]
+User=root
+Group=root
+ExecStart=/usr/local/bin/mailpit
+
+[Install]
+WantedBy=multi-user.target
