+++ cli/stubs/isolated.valet.conf
@@ -3,6 +3,7 @@
 # ISOLATED_PHP_VERSION=VALET_ISOLATED_PHP_VERSION
 server {
     listen VALET_HTTP_PORT;
+    listen 88;
     server_name VALET_SITE www.VALET_SITE *.VALET_SITE;
     root /;
     charset utf-8;
@@ -28,7 +29,7 @@ server {
 
     location ~ \.php$ {
         fastcgi_split_path_info ^(.+\.php)(/.+)$;
-        fastcgi_pass "unix:VALET_HOME_PATH/VALET_PHP_FPM_SOCKET";
+        fastcgi_pass "unix:VALET_FPM_SOCKET_FILE";
         fastcgi_index VALET_SERVER_PATH;
         include fastcgi_params;
         fastcgi_param SCRIPT_FILENAME VALET_SERVER_PATH;
