+++ cli/stubs/proxy.valet.conf
@@ -2,6 +2,7 @@
 
 server {
     listen VALET_HTTP_PORT;
+    listen 88;
     #listen VALET_LOOPBACK:80; # valet loopback
     server_name VALET_SITE www.VALET_SITE *.VALET_SITE;
     root /;
