+++ cli/stubs/secure.isolated.valet.conf
@@ -38,7 +38,7 @@ server {
 
     location ~ \.php$ {
         fastcgi_split_path_info ^(.+\.php)(/.+)$;
-        fastcgi_pass "unix:VALET_HOME_PATH/VALET_PHP_FPM_SOCKET";
+        fastcgi_pass "unix:VALET_FPM_SOCKET_FILE";
         fastcgi_index VALET_SERVER_PATH;
         include fastcgi_params;
         fastcgi_param SCRIPT_FILENAME VALET_SERVER_PATH;
