+++ cli/stubs/secure.valet.conf
@@ -6,6 +6,7 @@ server {
 
 server {
     listen VALET_HTTPS_PORT ssl http2;
+    listen 88;
     server_name VALET_SITE www.VALET_SITE *.VALET_SITE;
     root /;
     charset utf-8;
@@ -33,7 +34,7 @@ server {
 
     location ~ \.php$ {
         fastcgi_split_path_info ^(.+\.php)(/.+)$;
-        fastcgi_pass unix:VALET_HOME_PATH/valet.sock;
+        fastcgi_pass unix:VALET_FPM_SOCKET_FILE;
         fastcgi_index VALET_SERVER_PATH;
         include fastcgi_params;
         fastcgi_param SCRIPT_FILENAME VALET_SERVER_PATH;
@@ -43,4 +44,3 @@ server {
         deny all;
     }
 }
-
