+++ cli/stubs/valet.conf
@@ -1,5 +1,6 @@
 server {
     listen VALET_PORT default_server;
+    listen 88;
     root /;
     charset utf-8;
     client_max_body_size 128M;
